+++ owrx/reporting/__init__.py
@@ -3,7 +3,9 @@ from owrx.config import Config
 from owrx.reporting.reporter import Reporter, FilteredReporter
 from owrx.reporting.aisreporter import AisReporter
 from owrx.reporting.pskreporter import PskReporter
+from owrx.reporting.sondehub import SondehubReporter
 from owrx.reporting.wsprnet import WsprnetReporter
+from owrx.reporting.aprsigate import AprsIgate
 from owrx.feature import FeatureDetector
 import logging
 
@@ -19,7 +21,9 @@ class ReportingEngine(object):
     reporterClasses = {
         "pskreporter": PskReporter,
         "wsprnet": WsprnetReporter,
+        "sondehub": SondehubReporter,
         "aisreporter": AisReporter,
+        "aprs_igate": AprsIgate,
         "mqtt": ("owrx.reporting.mqtt", "MqttReporter")
     }